Refactor with common component.
