+++ assets/js/modules/ads/components/setup/index.js
@@ -16,7 +16,6 @@
  * limitations under the License.
  */
 
-export { default as SetupEnhancedConversionTrackingNotice } from './SetupEnhancedConversionTrackingNotice';
 export { default as SetupForm } from './SetupForm';
 export { default as SetupMain } from './SetupMain';
 export { default as SetupMainPAX } from './SetupMainPAX';
