+++ assets/js/modules/analytics-4/components/setup/index.js
@@ -16,7 +16,6 @@
  * limitations under the License.
  */
 
-export { default as SetupEnhancedConversionTrackingNotice } from './SetupEnhancedConversionTrackingNotice';
 export { default as SetupEnhancedMeasurementSwitch } from './SetupEnhancedMeasurementSwitch';
 export { default as SetupUseSnippetSwitch } from './SetupUseSnippetSwitch';
 export { default as SetupFormFields } from './SetupFormFields';
